[issue-33] Change category selection logic to intersect the results for different categories
